Split crl module in two parts
